+++ proxy/settings.js
@@ -1,9 +1,0 @@
-module.exports = {
-    dbhost: 'localhost',
-    dbuser: 'mageproxy',
-    dbpass: 'Proxypass_1234',
-    dbname: 'magentoproxy',
-
-    elasticuri: 'localhost:9200',
-    magentouri: 'http://localhost:90'
-};