+++ proxy/src/utils/Logger.ts
@@ -1,0 +1,7 @@
+import { createLogger, transports } from 'winston';
+
+export const Logger = createLogger({
+  transports: [
+    new transports.Console()
+  ]
+});